v5.8.2

+++ webfrontend/html/bin/check_on_state.php
@@ -14,29 +14,30 @@ $updatefile 	= "/run/shm/Sonos4lox_update.json";				// Status file during Sonos 
 	echo '<PRE>';
 	# check if script/Sonos Plugin is off
 	if (file_exists($off_file)) {
+		echo "<WARNING> Sonos Plugin is turned off. Please turn on".PHP_EOL;
 		exit;
 	}
 	# check if Sonos Firmware Update is running
 	if (file_exists($updatefile)) {
-		echo "<WARNING> Sonos Update is currently running, we abort here...".PHP_EOL;
+		echo "<ERROR> Sonos Update is currently running, we abort here...".PHP_EOL;
 		exit;
 	}
 
 	if (file_exists($lbpconfigdir . "/" . $configfile))    {
 		$config = json_decode(file_get_contents($lbpconfigdir . "/" . $configfile), TRUE);
 	} else {
-		#$params = [	"name" => "Sonos PHP",
-		#		"filename" => "$lbplogdir/sonos.log",
-		#		"append" => 1,
-		#		"addtime" => 1,
-		#		];
-		#$log = LBLog::newLog($params);
-		#LOGCRIT('Sonos: bin/check_on_state.php: The configuration file could not be loaded, the file may be disrupted. We have to abort :-(');
+		echo "<ERROR> The configuration file could not be loaded, the file may be disrupted. We have to abort :-(".PHP_EOL;
 		exit;
 	}
 	// Übernahme und Deklaration von Variablen aus der Konfiguration
 	$sonoszonen = $config['sonoszonen'];
 	#print_r($sonoszonen);
+	
+	# check if Online check is turned off in config
+	if ($config['SYSTEM']['checkonline'] === "false")   {
+		echo "<ERROR> Online check is not active".PHP_EOL;
+		exit(0);
+	}
 	
 	# check if folder exists
 	if (!is_dir("$lbpdatadir/PlayerStatus/"))   {
@@ -44,7 +45,7 @@ $updatefile 	= "/run/shm/Sonos4lox_update.json";				// Status file during Sonos 
 	}
 	
 	# Delete all Files
-	@array_map('unlink', glob("$lbpdatadir/PlayerStatus/s4lox_on_*.txt"));
+	#@array_map('unlink', glob("$lbpdatadir/PlayerStatus/s4lox_on_*.txt"));
 	
 	// prüft den Onlinestatus jeder Zone
 	#$zonesonline = array();
@@ -57,8 +58,10 @@ $updatefile 	= "/run/shm/Sonos4lox_update.json";				// Status file during Sonos 
 			#array_push($zonesonline, $zonen);
 			fclose($handle);
 			# create tmp file for each zone are online
-			file_put_contents("$lbpdatadir/PlayerStatus/s4lox_on_".$zonen.".txt", "on");
-			echo "<INFO> Player Online file 's4lox_on_".$zonen.".txt' has been created".PHP_EOL;
+			if (!file_exists("$lbpdatadir/PlayerStatus/s4lox_on_".$zonen.".txt"))   {
+				file_put_contents("$lbpdatadir/PlayerStatus/s4lox_on_".$zonen.".txt", "on");
+				echo "<OK> Player Online file 's4lox_on_".$zonen.".txt' has been created".PHP_EOL;
+			}
 			# get Player Mini PNG
 			$url = 'http://'.$ip[0].':1400/img/icon-'.$ip[7].'.png';
 			$img = $lbphtmldir.'/images/icon-'.$ip[7].'.png';
@@ -66,9 +69,14 @@ $updatefile 	= "/run/shm/Sonos4lox_update.json";				// Status file during Sonos 
 				file_put_contents($img, file_get_contents($url));
 			}
 		} else {
-			echo "<WARNING> File for Player '".$zonen."' has not been created. Maybe the Player is off... if so, please put Online and reboot Loxberry!".PHP_EOL;
+			if (file_exists("$lbpdatadir/PlayerStatus/s4lox_on_".$zonen.".txt")) {
+				unlink("$lbpdatadir/PlayerStatus/s4lox_on_".$zonen.".txt");
+				echo "<INFO> Player Online file 's4lox_on_".$zonen.".txt' has been deleted".PHP_EOL;
+			}
+			#echo "<WARNING> File for Player '".$zonen."' has not been created. Maybe the Player is off... if so, please put Online and reboot Loxberry!".PHP_EOL;
 			#LOGWARN("Sonos: bin/check_on_state.php: Zone $zonen seems to be Offline, please check your power/network settings");
 		}	
 	}
+	echo "<INFO> Online check for Players has been successful executed".PHP_EOL;
 	
 ?>
